Add posts and votes

+++ app/models/model_vote.go
@@ -16,5 +16,5 @@ type Vote struct {
 	Nickname string `json:"nickname"`
 
 	// Отданный голос.
-	Voice float32 `json:"voice"`
+	Voice int32 `json:"voice"`
 }
